chore: separate empty contents
